Make Content depend on ContentTemplate.
Make the `ContentPresenter.Content` property depend on `ContentPresenter.ContentTemplate` because otherwise a temporary control will be created by the presenter using the default template between the time `Content` and `ContentTemplate` is bound.

However, because Portable.Xaml doesn't actually _support_ `DependsOnAttribute` yet, also move the bindings for `ContentTemplate` above `Content` in the default themes.

This works around #1271 although #1284 fixes the underlying cause of the leak - this is more to prevent unnecessary controls being created.