Don't await updates in render loop.
The render loop should not be waiting for an update to occur on the UI thread before rendering a frame. Instead of awaiting the call, simply call `IDispatcher.Post` to fire-and-forget. Placed a guard around the update to make sure multiple updates don't get queued if an update doesn't complete in a single frame.

Also don't call `IRenderLoopTask.Update` unless `IRenderLoopTask.NeedsUpdate == true`.

Fixes #1920