Fix order of inherited property changed events.
Previously `PropertyChanged` was raised on the child before the parent because the inherited value change was being notified by listening to the parent `PropertyChanged` event, and this event handler was added first.

Added an `InheritablePropertyChanged` event which will be called only after all other property changed events have been raised and only for inheritable properties.