Don't UseLayoutRounding on ScrollBar.
Its default size of 18 with a thumb size of 8 can't be laid out to device pixels with a scaling of 1.5x. Previously, the thumb had more space on the right than the left, with this change it had more space on the left than the right. For some reason though it was more noticably offset with the space on the left. We may need to change the default sizes for `ScrollBar`.