Apply global clip to adorner element.
Previously, the adorner clip was not taking into account the clip inherited from ancestor controls, resulting in #3984.

Fix this by appling the clip from `TransformedBounds.Clip` by translating it to local coordinates first. Note that this doesn't work well for controls with a render transform applied, but at least it fixes the issue in the common case.

Fixes #3984