Correctly handle default values in batch update notifications.
We were previously calling `Optional.GetValueOrDefault<T>()` in order to convert an `Optional<object>` to an `Optional<T>` which meant that empty values were getting converted to default values, meaning that in certain cases change notifications weren't raised.

Add an `Optional<object>.Cast<T>()` methods (like we have in `BindingValue`) and use that in order to preserve empty values when ending a batch update operation.