Fix fetching neutral value in animations.
Animators fetched the neutral value when the animation is instantiated but did not update the neutral value if it's updated after that.

#6063 was caused by the fact that the neutral value does not take effect until the batch update due to styling has finished on the control, which is _after_ the animation has been instantiated.

Listen for changes on the property and if the change is not an animated value change, update the neutral value.

This isn't perfect because it won't react to changes to the neutral value while the animation is actually running and producing values because `PropertyChanged` events don't get fired for non-active values: for that we'd need to hook into `OnPropertyChangedCore`. But this at least fixes the simple case of the initial neutral value.