Handle FP errors in RoundLayoutValueUp.
If the `value` passed to `RoundLayoutValueUp` has a floating point precision error (e.g. 79.333333333333343 in the test case) then when it's multiplied by `dpiScale` and rounded up, it will be rounded up to the a value one greater than it should be.

Fix this by rounding the initial value to an arbitrary precision of 8 digits.

Fixes #8866