Refactored how we switch control themes.
Instead of simply wiping all control themes and styles that are applied to a control, we can now just remove the `ValueFrame`s which relate to the control theme that was changed.

To do this, added `ValueFrame.FramePriority` which encodes both the  `BindingPriority` and source of the frame (style, control theme, templated parent control theme).