Enable resizing during fullscreen transition.
macOS seems to tie resizing of the NSView inside the window to the resizable style mask of the window somehow. If we programmatically transition a non-resizable window to fullscreen, the inner NSView's size isn't changed, so we need to make the window resizable during the fullscreen transition.

Makes the final two failing `WindowState` integration tests pass.